\name{sharedvariants}
\alias{sharedvariants}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Shared variants matrix
}
\description{
  Generates a matrix of shared variant counts between each sample pair.
}
\usage{
  sharedvariants(deepseqmat)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{deepseqmat}{
    Deep sequence matrix generated by the \code{\link{deepseqmat}} function.
  }
}

\value{
Returns a matrix in which each entry details the number of shared intermediate frequency variants between each pair of samples.
}
\examples{
data(deepseq)
W <- deepseqmat(deepseq)
Y <- sharedvariants(W)
}